all distance data type changed from float to double

+++ FileIO_UI/MetroTunnelDB.cs
@@ -187,7 +187,7 @@ namespace libMetroTunnelDB
             Array.Copy(_a, 0, a, 0, floatArrLength);
         }
 
-        public DataConv(int _RecordID, float _Distance)
+        public DataConv(int _RecordID, double _Distance)
         {
             RecordID = _RecordID;
             Distance = _Distance;
@@ -256,12 +256,12 @@ namespace libMetroTunnelDB
     public class ImageDisp
     {
         public int RecordID;
-        public float Distance;
+        public double Distance;
         public String[] FileUrl;
 
         public const int StringArrLength = 8;
 
-        public ImageDisp(int _RecordID, float _Distance, String[] _FileUrl)
+        public ImageDisp(int _RecordID, double _Distance, String[] _FileUrl)
         {
             RecordID = _RecordID;
             Distance = _Distance;
@@ -270,7 +270,7 @@ namespace libMetroTunnelDB
                 FileUrl[i] = _FileUrl[i];
         }
 
-        public ImageDisp(int _RecordID, float _Distance)
+        public ImageDisp(int _RecordID, double _Distance)
         {
             RecordID = _RecordID;
             Distance = _Distance;
@@ -289,7 +289,7 @@ namespace libMetroTunnelDB
         public int RecordID;
         public int CameraID;
         public int TimeStamp;
-        public float Distance;
+        public double Distance;
         public float[] x;
         public float[] y;
         public float[] s;
@@ -865,7 +865,7 @@ namespace libMetroTunnelDB
             DataConv entry = new DataConv
             (
                 reader.GetInt32("RecordID"),
-                reader.GetFloat("Distance"),
+                reader.GetDouble("Distance"),
                 DataEntry.ByteArr2FloatArr(s),
                 DataEntry.ByteArr2FloatArr(a)
             );
@@ -877,7 +877,7 @@ namespace libMetroTunnelDB
             DataDisp entry = new DataDisp
             (
                 reader.GetInt32("RecordID"),
-                reader.GetFloat("Distance"),
+                reader.GetDouble("Distance"),
                 reader.GetString("JsonString")
             );
             return entry;
@@ -889,7 +889,7 @@ namespace libMetroTunnelDB
             (
                 reader.GetInt32("RecordID"),
                 reader.GetInt32("TimeStamp"),
-                reader.GetFloat("Distance")
+                reader.GetDouble("Distance")
             );
             return entry;
         }
@@ -914,7 +914,7 @@ namespace libMetroTunnelDB
             ImageDisp entry = new ImageDisp
             (
                 reader.GetInt32("RecordID"),
-                reader.GetFloat("Distance"),
+                reader.GetDouble("Distance"),
                 arr
             );
             return entry;
@@ -1086,7 +1086,7 @@ namespace libMetroTunnelDB
             DoDelete(deleteDetectRecordStr);
         }
 
-        public void QueryDataOverview(ref List<DataOverview> arr, int RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataOverview(ref List<DataOverview> arr, int RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataOverview WHERE RecordID={0} AND Distance>={1} AND Distance<={2}";
             String queryStr = String.Format(formatStr, RecordID, min_Distance, max_Distance);
@@ -1100,7 +1100,7 @@ namespace libMetroTunnelDB
             DoQuery(queryStr, ref arr, ReadDataOverview);
         }
 
-        public void QueryDataOverview(ref List<DataOverview> arr, int min_RecordID, int max_RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataOverview(ref List<DataOverview> arr, int min_RecordID, int max_RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataOverview WHERE RecordID>={0} AND RecordID<={1} AND Distance>={2} AND Distance<={3}";
             String queryStr = String.Format(formatStr, min_RecordID, max_RecordID, min_Distance, max_Distance);
@@ -1120,21 +1120,21 @@ namespace libMetroTunnelDB
             DoQuery(queryStr, ref arr, ReadDataRaw);
         }
 
-        public void QueryDataConv(ref List<DataConv> arr, int RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataConv(ref List<DataConv> arr, int RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataConv WHERE RecordID={0} AND Distance>={1} AND Distance<={2}";
             String queryStr = String.Format(formatStr, RecordID, min_Distance, max_Distance);
             DoQuery(queryStr, ref arr, ReadDataConv);
         }
 
-        public void QueryDataConv(ref List<DataConv> arr, int min_RecordID, int max_RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataConv(ref List<DataConv> arr, int min_RecordID, int max_RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataConv WHERE RecordID>={0} AND RecordID<{1} AND Distance>={2} AND Distance<={3}";
             String queryStr = String.Format(formatStr, min_RecordID, max_RecordID, min_Distance, max_Distance);
             DoQuery(queryStr, ref arr, ReadDataConv);
         }
 
-        public void QueryDataConv(ref List<DataConv> arr, int RecordID, int QueryFrom, int maxQuery, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataConv(ref List<DataConv> arr, int RecordID, int QueryFrom, int maxQuery, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataConv WHERE RecordID={0} AND Distance>={1} AND Distance<={2} LIMIT {3},{4}";
             String queryStr = String.Format(formatStr, RecordID, min_Distance, max_Distance, QueryFrom, maxQuery);
@@ -1142,14 +1142,14 @@ namespace libMetroTunnelDB
         }
 
 
-        public void QueryDataDisp(ref List<DataDisp> arr, int RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataDisp(ref List<DataDisp> arr, int RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataDisp WHERE RecordID={0} AND Distance>={1} AND Distance<={2}";
             String queryStr = String.Format(formatStr, RecordID, min_Distance, max_Distance);
             DoQuery(queryStr, ref arr, ReadDataDisp);
         }
 
-        public void QueryDataDisp(ref List<DataDisp> arr, int min_RecordID, int max_RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryDataDisp(ref List<DataDisp> arr, int min_RecordID, int max_RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM DataDisp WHERE RecordID>={0} AND RecordID<={1} AND Distance>={2} AND Distance<={3}";
             String queryStr = String.Format(formatStr, min_RecordID, max_RecordID, min_Distance, max_Distance);
@@ -1184,7 +1184,7 @@ namespace libMetroTunnelDB
             DoQuery(queryStr, ref arr, ReadImageRaw);
         }
 
-        public void QueryImageDisp(ref List<ImageDisp> arr, int RecordID, float min_Distance = 0, float max_Distance = float.MaxValue)
+        public void QueryImageDisp(ref List<ImageDisp> arr, int RecordID, double min_Distance = 0, double max_Distance = double.MaxValue)
         {
             String formatStr = "SELECT * FROM ImageDisp WHERE RecordID={0} AND Distance>={1} AND Distance<={2}";
             String queryStr = String.Format(formatStr, RecordID, min_Distance, max_Distance);
